feat(meals): Refactor file structure and update imports for Meals feature
