+++ lib/features/meals/presentation/view/widgets/meals_body.dart
@@ -1,15 +1,15 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:fitness_app/core/utils/l10n/locale_keys.g.dart';
 import 'package:fitness_app/domain/meals/entity/category_entity.dart';
-import 'package:fitness_app/features/meals/view_model/meals_cubit.dart';
-import 'package:fitness_app/features/meals/view_model/meals_state.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:skeletonizer/skeletonizer.dart';
 
-import '../../../../core/assets/app_colors.dart';
-import '../../../../core/base/base_state.dart';
-import '../../../../core/utils/shared_widgets/custom_tab_bar.dart';
+import '../../../../../core/assets/app_colors.dart';
+import '../../../../../core/base/base_state.dart';
+import '../../../../../core/utils/shared_widgets/custom_tab_bar.dart';
+import '../../view_model/meals_cubit.dart';
+import '../../view_model/meals_state.dart';
 import 'meals_grid.dart';
 
 class MealsBody extends StatelessWidget {
