+++ lib/features/meals/presentation/view_model/meals_state.dart
